Runtime Type System & Projection fixes
***
Implemented runtime type checking for entities. This implementation now enables the developer to check if the entity has a such component of that type.

***
Projection matrix for the camera was not calculating the window size previously. Therefore, any window aspect ratio change caused due to window resizing really messed how the objects was being drawn to the screen. Entire render process is now based on the window size to better fit the aspect ratio