Initial ECS implementation finished
ECS is implemented in an isolated environment. It is currently not integrated into the engine. I am currently going through with extensive testing before replacing the current Entity-Component model with ECS. I want to keep this in a separate branch as  possible before merging it because I want to keep tests as isolated as possible in case any fuck up happens.